requirements.txt
flask-sqlalchemy
flask-wtf
flask
flask-login
gunicorn
python-dotenv
email-validator
Flask-MySQL
mysqlclient